+++ TagsCloud/TagsCloudDrawer/ImageSaveService/IImageSavior.cs
@@ -1,0 +1,9 @@
+﻿using System.Drawing;
+
+namespace TagsCloudDrawer.ImageSaveService
+{
+    public interface IImageSavior
+    {
+        void Save(string filename, Image image);
+    }
+}
