+++ TagsCloud/TagsCloudDrawer/ImageSaveService/PngSavior.cs
@@ -1,0 +1,14 @@
+﻿using System.Drawing;
+using System.Drawing.Imaging;
+using System.IO;
+
+namespace TagsCloudDrawer.ImageSaveService
+{
+    public class PngSavior : IImageSavior
+    {
+        public void Save(string filename, Image image)
+        {
+            image.Save(Path.ChangeExtension(filename, "png"), ImageFormat.Png);
+        }
+    }
+}
