--- TagsCloud/TagsCloudDrawer/ImageSavior/IImageSavior.cs
@@ -1,9 +1,0 @@
-﻿using System.Drawing;
-
-namespace TagsCloudDrawer.ImageSavior
-{
-    public interface IImageSavior
-    {
-        void Save(string filename, Image image);
-    }
-}
