--- TagsCloud/TagsCloudDrawer/ImageSavior/PngSavior.cs
@@ -1,14 +1,0 @@
-﻿using System.Drawing;
-using System.Drawing.Imaging;
-using System.IO;
-
-namespace TagsCloudDrawer.ImageSavior
-{
-    public class PngSavior : IImageSavior
-    {
-        public void Save(string filename, Image image)
-        {
-            image.Save(Path.ChangeExtension(filename, "png"), ImageFormat.Png);
-        }
-    }
-}
